+++ igvplot/__init__.py
@@ -57,7 +57,7 @@ from .view import (
 )
 from . import plot
 
-__version__ = "0.1.0"
+__version__ = "0.0.1"
 
 __all__ = [
     "Region",